fix(synthetic-shadow): do not patch lwc root elements
when using collections
