+++ packages/@lwc/synthetic-shadow/src/faux-shadow/traverse.ts
@@ -11,6 +11,7 @@ import {
     parentNodeGetter,
     compareDocumentPosition,
     DOCUMENT_POSITION_CONTAINS,
+    parentElementGetter,
 } from '../env/node';
 import { querySelectorAll } from '../env/element';
 import {
@@ -21,6 +22,85 @@ import {
     getShadowRoot,
 } from './shadow-root';
 import { arrayFromCollection } from '../shared/utils';
+
+// when finding a slot in the DOM, we can fold it if it is contained
+// inside another slot.
+function foldSlotElement(slot: HTMLElement) {
+    let parent = parentElementGetter.call(slot);
+    while (!isNull(parent) && isSlotElement(parent)) {
+        slot = parent as HTMLElement;
+        parent = parentElementGetter.call(slot);
+    }
+    return slot;
+}
+
+function isNodeSlotted(host: Element, node: Node): boolean {
+    if (process.env.NODE_ENV !== 'production') {
+        assert.invariant(
+            host instanceof HTMLElement,
+            `isNodeSlotted() should be called with a host as the first argument instead of ${host}`
+        );
+        assert.invariant(
+            node instanceof Node,
+            `isNodeSlotted() should be called with a node as the second argument instead of ${node}`
+        );
+        assert.invariant(
+            compareDocumentPosition.call(node, host) & DOCUMENT_POSITION_CONTAINS,
+            `isNodeSlotted() should never be called with a node that is not a child node of ${host}`
+        );
+    }
+    const hostKey = getNodeKey(host);
+    // this routine assumes that the node is coming from a different shadow (it is not owned by the host)
+    // just in case the provided node is not an element
+    let currentElement = node instanceof Element ? node : parentElementGetter.call(node);
+    while (!isNull(currentElement) && currentElement !== host) {
+        const elmOwnerKey = getNodeNearestOwnerKey(currentElement);
+        const parent = parentElementGetter.call(currentElement);
+        if (elmOwnerKey === hostKey) {
+            // we have reached an element inside the host's template, and only if
+            // that element is an slot, then the node is considered slotted
+            return isSlotElement(currentElement);
+        } else if (parent === host) {
+            return false;
+        } else if (!isNull(parent) && getNodeNearestOwnerKey(parent) !== elmOwnerKey) {
+            // we are crossing a boundary of some sort since the elm and its parent
+            // have different owner key. for slotted elements, this is possible
+            // if the parent happens to be a slot.
+            if (isSlotElement(parent)) {
+                /**
+                 * the slot parent might be allocated inside another slot, think of:
+                 * <x-root> (<--- root element)
+                 *    <x-parent> (<--- own by x-root)
+                 *       <x-child> (<--- own by x-root)
+                 *           <slot> (<--- own by x-child)
+                 *               <slot> (<--- own by x-parent)
+                 *                  <div> (<--- own by x-root)
+                 *
+                 * while checking if x-parent has the div slotted, we need to traverse
+                 * up, but when finding the first slot, we skip that one in favor of the
+                 * most outer slot parent before jumping into its corresponding host.
+                 */
+                currentElement = getNodeOwner(foldSlotElement(parent as HTMLElement));
+                if (!isNull(currentElement)) {
+                    if (currentElement === host) {
+                        // the slot element is a top level element inside the shadow
+                        // of a host that was allocated into host in question
+                        return true;
+                    } else if (getNodeNearestOwnerKey(currentElement) === hostKey) {
+                        // the slot element is an element inside the shadow
+                        // of a host that was allocated into host in question
+                        return true;
+                    }
+                }
+            } else {
+                return false;
+            }
+        } else {
+            currentElement = parent;
+        }
+    }
+    return false;
+}
 
 export function getNodeOwner(node: Node): HTMLElement | null {
     if (!(node instanceof Node)) {
@@ -68,6 +148,30 @@ export function isNodeOwnedBy(owner: Element, node: Node): boolean {
 export function shadowRootChildNodes(root: SyntheticShadowRootInterface): Array<Element & Node> {
     const elm = getHost(root);
     return getAllMatches(elm, arrayFromCollection(childNodesGetter.call(elm)));
+}
+
+export function getAllSlottedMatches(
+    host: Element,
+    nodeList: NodeList | Node[]
+): Array<Node & Element> {
+    const filteredAndPatched = [];
+    for (let i = 0, len = nodeList.length; i < len; i += 1) {
+        const node = nodeList[i];
+        if (!isNodeOwnedBy(host, node) && isNodeSlotted(host, node)) {
+            ArrayPush.call(filteredAndPatched, node);
+        }
+    }
+    return filteredAndPatched;
+}
+
+export function getFirstSlottedMatch(host: Element, nodeList: Element[]): Element | null {
+    for (let i = 0, len = nodeList.length; i < len; i += 1) {
+        const node = nodeList[i] as Element;
+        if (!isNodeOwnedBy(host, node) && isNodeSlotted(host, node)) {
+            return node;
+        }
+    }
+    return null;
 }
 
 export function getAllMatches(owner: Element, nodeList: Node[]): Array<Element & Node> {
